requirements.txt
Django==4.2.7
djangorestframework==3.14.0
langchain==0.1.0
langchain-google-genai==0.0.11
langchain-community==0.0.10
python-decouple==3.8